pyproject.toml
[tool.poetry]
name = "onlineshop"
version = "0.1.0"
description = ""
authors = ["Vladislav <ilusevicv@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.115.8"}
SQLAlchemy = "^2.0.38"
alembic = "^1.14.1"
asyncpg = "^0.30.0"
python-dotenv = "^1.0.1"
python-jose = "^3.4.0"
passlib = "^1.7.4"
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
